Lots of changes. Tidied up front end, applied more logic to login process and it now pulls clients back from the table correctly.

+++ PersonalTrainer/Models/MoreClientData.cs
@@ -14,18 +14,14 @@ namespace PersonalTrainer.Models
         [Display(Name = "Body Composition")]
         public string BodyComposition { get; set; }
 
-
-        // already collected in workout.cs
         [Display(Name ="Body Weight")]
-        public double BodyWeight { get; set; }
+        public decimal BodyWeight { get; set; }
 
         [Display(Name = "Body Fat %")]
         public double BodyFatPercentage { get; set; }
 
-
-        // already collected in workout.cs
         [Display(Name = "Height")]
-        public double Height { get; set; }
+        public decimal Height { get; set; }
 
         [Display(Name = "Waist")]
         public int Waist { get; set; }
